+++ NEWS.md
@@ -45,6 +45,7 @@
 * Rename internal modules to follow golem standard. I.E. 'mod_...'
 * Fixed bug where the logging file was not being set
 * Source riskmetric from GitHub
+* Added decision automation capabilities where the user can set decision rules for uploaded packages to be auto-assigned
 
 # riskassessment 0.0.1
 * Initiated simple `app.R` for easier deployment using `runURL("https://github.com/pharmaR/riskassessment/archive/master.zip")` and `shiny::runGitHub('riskassessment', 'pharmaR')`